fixed schema and made toHex more compact

+++ js/mtproto/skeletonSchema.js
@@ -1,3 +1,3 @@
 export default `
-  notLoggedSkeleton# auth_key_id:int32 message_id:int32 message_data_length:int16 message_data:bytes = _;
+  notLoggedSkeleton# auth_key_id:int64 message_id:int64 message_data_length:int32 message_data:bytes = _;
 `
